+++ cpp/include/tensorrt_llm/batch_manager/createNewDecoderRequests.h
@@ -17,11 +17,15 @@
 
 #pragma once
 
+#include "tensorrt_llm/batch_manager/common.h"
 #include "tensorrt_llm/common/algorithm.h"
+#include "tensorrt_llm/common/optionalRef.h"
+#include "tensorrt_llm/runtime/bufferManager.h"
 #include "tensorrt_llm/runtime/common.h"
 #include "tensorrt_llm/runtime/iTensor.h"
 #include "tensorrt_llm/runtime/modelConfig.h"
 #include "tensorrt_llm/runtime/request.h"
+#include "tensorrt_llm/runtime/worldConfig.h"
 
 namespace tensorrt_llm::runtime
 {
@@ -40,6 +44,8 @@ class DecoderState;
 
 namespace tensorrt_llm::batch_manager
 {
+class RuntimeBuffers;
+class DecoderInputBuffers;
 
 class CreateNewDecoderRequests : Algorithm
 {
@@ -55,12 +61,22 @@ public:
     using DecodingOutput = runtime::DecodingOutput;
     using SpeculativeDecodingMode = runtime::SpeculativeDecodingMode;
     using GptDecoderBatched = runtime::GptDecoderBatched;
+    template <typename T>
+    using OptionalRef = tensorrt_llm::common::OptionalRef<T>;
 
-    CreateNewDecoderRequests() = default;
+    CreateNewDecoderRequests(bool speculativeDecodingFastLogits, bool isLeaderInOrchMode, bool isNormalizeLogProbs)
+        : mSpeculativeDecodingFastLogits(speculativeDecodingFastLogits)
+        , mIsLeaderInOrchMode(isLeaderInOrchMode)
+        , mIsNormalizeLogProbs(isNormalizeLogProbs)
+    {
+    }
 
-    void operator()(TensorPtr const& batchSlots, std::vector<runtime::decoder_batch::Request> const& requests,
-        std::vector<SamplingConfig> const& samplingConfigs, runtime::ModelConfig const& modelConfig,
-        GptDecoderBatched& decoder, CudaStream const& runtimeStream, SizeType32 maxSequenceLength) const;
+    std::tuple<TensorPtr, std::vector<runtime::decoder_batch::Request>, std::vector<runtime::SamplingConfig>>
+    operator()(runtime::ModelConfig const& modelConfig, runtime::WorldConfig const& worldConfig,
+        executor::DecodingConfig const& decodingConfig, RequestVector const& contextRequests,
+        runtime::BufferManager const& bufferManager, nvinfer1::DataType logitsType,
+        DecoderInputBuffers& inputBuffers, GptDecoderBatched& decoder, CudaStream const& runtimeStream,
+        SizeType32 maxSequenceLength, SizeType32 beamWidth, OptionalRef<RuntimeBuffers const> buffers) const;
 
     //! @brief Initialize the decoder at `batchSlot` with a new `request`. Exposed only for static batching via
     //! GptDecoderBatched::newBatch()
@@ -96,6 +112,27 @@ private:
     //! @brief Setups decoder internal tensors for new Eagle request
     static void newRequestEagle(SizeType32 batchIdx, runtime::decoder_batch::Request const& request,
         runtime::ModelConfig const& modelConfig, DecodingOutput& jointDecodingOutput, CudaStream const& runtimeStream);
+
+    [[nodiscard]] std::tuple<TensorPtr, std::vector<runtime::decoder_batch::Request>, std::vector<runtime::SamplingConfig>>
+    generateRequestOptions(runtime::ModelConfig const& modelConfig, runtime::WorldConfig const& worldConfig,
+        executor::DecodingConfig const& decodingConfig, RequestVector const& contextRequests,
+        runtime::BufferManager const& bufferManager, nvinfer1::DataType logitsType, DecoderInputBuffers& inputBuffers,
+        runtime::decoder::DecoderState& decoderState, SizeType32 beamWidth, runtime::CudaStream const& stream,
+        OptionalRef<RuntimeBuffers const> buffers = std::nullopt) const;
+
+    [[nodiscard]] std::vector<runtime::decoder_batch::Request> createDecoderRequests(
+        RequestVector const& finishedContextRequests, TensorPtr const& inputIds,
+        executor::DecodingConfig const& decodingConfig, runtime::BufferManager const& bufferManager,
+        nvinfer1::DataType logitsType, runtime::ModelConfig const& modelConfig, runtime::WorldConfig const& worldConfig,
+        OptionalRef<RuntimeBuffers const> buffers) const;
+
+    [[nodiscard]] std::shared_ptr<runtime::ITensor> retrieveDraftLogits(runtime::ModelConfig const& modelConfig,
+        runtime::WorldConfig const& worldConfig, std::shared_ptr<runtime::ITensor> const& tensor,
+        runtime::BufferManager const& bufferManager) const;
+
+    bool mSpeculativeDecodingFastLogits;
+    bool mIsLeaderInOrchMode;
+    bool mIsNormalizeLogProbs;
 };
 
 } // namespace tensorrt_llm::batch_manager
